watch_buffer_fill
Function get_buff_size

pub(crate) fn get_buff_size(
    which: &RamBuffer,
    buff_start: &mut u32,
) -> Result<u32, RegisterError>

! Get the blob buffer size from occupancy register

Read out the occupancy register and compare to a previously recoreded value. Everything is u32 (the register can’t hold more)

The size of the buffer can only be defined compared to a start value. If the value rools over, the size then does not make no longer sense and needs to be updated.

#Arguments: